requirements.txt
streamlit
crewai
crewai-tools
langchain-openai
python-dotenv